// src/shared/interfaces/users.interface.ts
export interface ITokenPayload {
  id: number;
  user_name: string;
  email: string;
  iat?: number;
  exp?: number;
}

export interface IUser {
  id?: number;
  user_name: string;
  email: string;
  password: string;
  full_name: string;
  date_of_birth: string;
  avatar_image_url: string;
  reset_password_token?: string | null;
  reset_password_expires?: Date | null;
}

export interface IRegisterUser {
  user_name: string;
  email: string;
  password: string;
  confirm_password: string;
  full_name: string;
  date_of_birth: string;
  avatar_image_url: string;
}